fixbug phiếu nhập sách

+++ GUI/fThemChiTietPhieuNhapSach.cs
@@ -15,6 +15,7 @@ namespace GUI
 {
     public partial class fThemChiTietPhieuNhapSach : Form
     {
+        private int DaLuu; 
         public fThemChiTietPhieuNhapSach()
         {
             InitializeComponent();
@@ -22,50 +23,76 @@ namespace GUI
 
         private void fThemChiTietPhieuNhapSach_Load(object sender, EventArgs e)
         {
-            
+            DaLuu = 0;
             cboMaSach.DataSource = SachBUS.Instance.GetAllSach();
             cboMaSach.DisplayMember = "MASACH";
-            cboMaSach.ValueMember = "TENSACH";
+            cboMaSach.ValueMember = "MASACH";
             cboMaSach.SelectedIndex = -1;
             txtTenSach.Enabled = false;
             txtTenSach.Text = "";
-            txtDonGia.Text = "0";
+            txtThanhTien.Text = "0";
             txtSoPhieuNhap.Enabled = false;
             txtTongTien.Enabled = false;
             txtThanhTien.Enabled = false;
+            txtDonGia.Enabled = false;
+            dtmNgayLap.Enabled = false;
             txtSoPhieuNhap.Text = PhieuNhapSachBUS.Instance.GetLatestPhieuNhapSach().Rows[0][1].ToString();
             txtTongTien.Text = "0";
+            dgvDSSachNhap.AutoSizeColumnsMode = DataGridViewAutoSizeColumnsMode.Fill;
+            dgvDSSachNhap.DataSource = CT_PhieuNhapSachBUS.Instance.GetCT_PhieuNhapSachByIDPNS(txtSoPhieuNhap.Text.Replace("PNS", "000"));
+            dgvDSSachNhap.Columns[0].HeaderCell.Value = "Mã sách";
+            dgvDSSachNhap.Columns[1].HeaderCell.Value = "Tên sách";
+            dgvDSSachNhap.Columns[2].HeaderCell.Value = "Số lượng";
+            dgvDSSachNhap.Columns[3].HeaderCell.Value = "Đơn giá";
+            dgvDSSachNhap.Columns[4].HeaderCell.Value = "Thành tiền";
+
         }
 
         private void cboMaSach_SelectedIndexChanged(object sender, EventArgs e)
         {
             if (cboMaSach.SelectedValue != null)
             {
-                txtTenSach.Text = cboMaSach.SelectedValue.ToString();
+                DataTable data =SachBUS.Instance.GetAllSachByMaSach(cboMaSach.SelectedValue.ToString());
+                if( data.Rows.Count > 0)
+                {
+                    txtTenSach.Text = data.Rows[0][1].ToString();
+                    txtDonGia.Text = data.Rows[0][6].ToString();
+                    txtThanhTien.Text = Convert.ToString(Convert.ToInt32(txtDonGia.Text.ToString()) * Convert.ToInt32(domSoLuong.Text.ToString()));
+                    domSoLuong.Text = "1";
+                }      
+            }
+            else
+            {
+                txtTenSach.Text = "";
+                txtDonGia.Text = "0";
             }    
-        }
-
-        private void txtDonGia_TextChanged(object sender, EventArgs e)
-        {
-            txtThanhTien.Text = Convert.ToString(Convert.ToInt32(txtDonGia.Text.ToString()) * Convert.ToInt32(domSoLuong.Text.ToString()));
-        }
-
-        private void domSoLuong_SelectedItemChanged(object sender, EventArgs e)
-        {
-            txtThanhTien.Text = Convert.ToString(Convert.ToInt32(txtDonGia.Text.ToString()) * Convert.ToInt32(domSoLuong.Text.ToString()));
         }
 
         private void btnThemSachNhap_Click(object sender, EventArgs e)
         {
-            CT_PhieuNhapSach ct_PNS = new CT_PhieuNhapSach();
-            ct_PNS.IDPNS = Convert.ToInt32(txtSoPhieuNhap.Text.Replace("PNS", "000"));
-            ct_PNS.IDSach = Convert.ToInt32(cboMaSach.Text.Replace("SA", "00"));
-            ct_PNS.DonGia = Convert.ToInt32(txtDonGia.Text);
-            ct_PNS.SoLuongNhap = Convert.ToInt32(domSoLuong.Text);
-            ct_PNS.ThanhTien = (Convert.ToInt32(txtThanhTien.Text.ToString()));
-            CT_PhieuNhapSachBUS.Instance.InsertCT_PhieuNhapSach(ct_PNS);
-            dgvDSSachNhap.DataSource = CT_PhieuNhapSachBUS.Instance.GetCT_PhieuNhapSachByIDPNS(ct_PNS.IDPNS.ToString());
-            txtTongTien.Text = ((Convert.ToInt32(txtTongTien.Text.ToString())) + ct_PNS.ThanhTien).ToString();
+            if (cboMaSach.SelectedValue == null)
+            {
+                MessageBox.Show("Vui lòng chọn lại mã sách!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                cboMaSach.Focus();
+            }
+            else if (CT_PhieuNhapSachBUS.Instance.GetCT_PhieuNhapSachByIDPNSAndIDSach(txtSoPhieuNhap.Text.Replace("PNS", "000"), cboMaSach.Text.Replace("SA", "00")).Rows.Count > 0)
+            {
+                MessageBox.Show("Đã tồn tại chi tiết phiếu nhập. Vui lòng kiếm tra lại!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                cboMaSach.Focus();
+            }
+            else
+            {
+                CT_PhieuNhapSach ct_PNS = new CT_PhieuNhapSach();
+                ct_PNS.IDPNS = Convert.ToInt32(txtSoPhieuNhap.Text.Replace("PNS", "000"));
+                ct_PNS.IDSach = Convert.ToInt32(cboMaSach.Text.Replace("SA", "00"));
+                ct_PNS.DonGia = Convert.ToInt32(txtDonGia.Text);
+                ct_PNS.SoLuongNhap = Convert.ToInt32(domSoLuong.Text);
+                ct_PNS.ThanhTien = (Convert.ToInt32(txtThanhTien.Text.ToString()));
+                CT_PhieuNhapSachBUS.Instance.InsertCT_PhieuNhapSach(ct_PNS);
+                dgvDSSachNhap.DataSource = CT_PhieuNhapSachBUS.Instance.GetCT_PhieuNhapSachByIDPNS(ct_PNS.IDPNS.ToString());
+                txtTongTien.Text = ((Convert.ToInt32(txtTongTien.Text.ToString())) + ct_PNS.ThanhTien).ToString();
+            }    
+            
         }
 
         private void cboSoPhieuNhap_SelectedIndexChanged(object sender, EventArgs e)
@@ -96,40 +123,75 @@ namespace GUI
 
         private void btnXoaSachNhap_Click(object sender, EventArgs e)
         {
-            CT_PhieuNhapSachBUS.Instance.DeleteCT_PhieuNhapSach(txtSoPhieuNhap.Text.Replace("PNS", "000"), cboMaSach.Text.Replace("SA", "00"));
-            dgvDSSachNhap.DataSource = CT_PhieuNhapSachBUS.Instance.GetCT_PhieuNhapSachByIDPNS(txtSoPhieuNhap.Text.Replace("PNS", "000"));
-            cboMaSach.Text = "";
-            txtTenSach.Text = "";
-            txtDonGia.Text = "0";
-            txtThanhTien.Text = "0";
-            domSoLuong.Text = "00";
-            int TongTien = 0;
-            for (int i = 0; i < dgvDSSachNhap.Rows.Count - 1; i++)
+            if (cboMaSach.SelectedValue == null)
             {
-                TongTien = TongTien + Convert.ToInt32(dgvDSSachNhap.Rows[i].Cells[4].Value);
+                MessageBox.Show("Vui lòng chọn lại mã sách!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                cboMaSach.Focus();
             }
-            txtTongTien.Text = TongTien.ToString();
+            else if (CT_PhieuNhapSachBUS.Instance.GetCT_PhieuNhapSachByIDPNSAndIDSach(txtSoPhieuNhap.Text.Replace("PNS", "000"), cboMaSach.Text.Replace("SA", "00")).Rows.Count > 0)
+            {
+                CT_PhieuNhapSachBUS.Instance.DeleteCT_PhieuNhapSach(txtSoPhieuNhap.Text.Replace("PNS", "000"), cboMaSach.Text.Replace("SA", "00"));
+                dgvDSSachNhap.DataSource = CT_PhieuNhapSachBUS.Instance.GetCT_PhieuNhapSachByIDPNS(txtSoPhieuNhap.Text.Replace("PNS", "000"));
+                txtTongTien.Text = (Convert.ToInt32(txtTongTien.Text) - Convert.ToInt32(txtThanhTien.Text)).ToString();
+                cboMaSach.Text = "";
+                txtTenSach.Text = "";
+                txtDonGia.Text = "0";
+                txtThanhTien.Text = "0";
+                domSoLuong.Text = "1";    
+            }
+            else
+            {
+                MessageBox.Show("Không tồn tại chi tiết phiếu nhập. Vui lòng kiếm tra lại!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                cboMaSach.Focus();
+            }    
+            
         }
 
         private void btnLuu_Click(object sender, EventArgs e)
         {
-            PhieuNhapSach PNS = new PhieuNhapSach();
-            PNS.SoPNS = txtSoPhieuNhap.Text;
-            PNS.TongTien = Convert.ToInt32(txtTongTien.Text.ToString());
-            PhieuNhapSachBUS.Instance.UpdatePhieuNhapSach(PNS);
-            CuonSach cuonsach = new CuonSach();
-            for (int i = 0; i < dgvDSSachNhap.Rows.Count - 1; i++)
+
+            if (CT_PhieuNhapSachBUS.Instance.GetCT_PhieuNhapSachByIDPNS(txtSoPhieuNhap.Text.Replace("PNS", "000")).Rows.Count > 0)
             {
-                for(int j = 0; j < Convert.ToInt32(dgvDSSachNhap.Rows[i].Cells[2].Value.ToString()); j++)
-                    CuonSachBUS.Instance.InsertCuonSach(dgvDSSachNhap.Rows[i].Cells[0].Value.ToString().Replace("SA", "00"));//0 ma sah//2 la so luong//3 don gia
-                SachBUS.Instance.UpdateSach(Convert.ToInt32(dgvDSSachNhap.Rows[i].Cells[2].Value.ToString()), dgvDSSachNhap.Rows[i].Cells[3].Value.ToString(), dgvDSSachNhap.Rows[i].Cells[0].Value.ToString());
-                
+                DaLuu = 1;
+                PhieuNhapSach PNS = new PhieuNhapSach();
+                PNS.SoPNS = txtSoPhieuNhap.Text;
+                PNS.TongTien = Convert.ToInt32(txtTongTien.Text.ToString());
+                PhieuNhapSachBUS.Instance.UpdatePhieuNhapSach(PNS);
+                CuonSach cuonsach = new CuonSach();
+                for (int i = 0; i < dgvDSSachNhap.Rows.Count - 1; i++)
+                {
+                    for (int j = 0; j < Convert.ToInt32(dgvDSSachNhap.Rows[i].Cells[2].Value.ToString()); j++)
+                        CuonSachBUS.Instance.InsertCuonSach(dgvDSSachNhap.Rows[i].Cells[0].Value.ToString().Replace("SA", "00"));
+                    SachBUS.Instance.UpdateSach(Convert.ToInt32(dgvDSSachNhap.Rows[i].Cells[2].Value.ToString()), dgvDSSachNhap.Rows[i].Cells[3].Value.ToString(), dgvDSSachNhap.Rows[i].Cells[0].Value.ToString());
+                }
+                MessageBox.Show("Thêm phiếu nhập sách thành công!", "Thông tin", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                this.Close();
             }
+            else 
+            {
+                MessageBox.Show("Không tồn tại sách nhập. Vui lòng kiếm tra lại!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                cboMaSach.Focus();
+            } 
+                
+            
         }
 
         private void btnHuy_Click(object sender, EventArgs e)
         {
-            this.Close();
+            if (DaLuu != 0)
+            {
+                this.Close();
+            }
+            else
+            {
+                PhieuNhapSachBUS.Instance.DeletePhieuNhapSach(txtSoPhieuNhap.Text);
+                this.Close();
+            }
+        }
+
+        private void domSoLuong_ValueChanged(object sender, EventArgs e)
+        {
+            txtThanhTien.Text = Convert.ToString(Convert.ToInt32(txtDonGia.Text.ToString()) * Convert.ToInt32(domSoLuong.Text.ToString()));
         }
     }
 }
